+++ extensions/bitwarden/src/constants/general.ts
@@ -2,6 +2,8 @@
 
 export const DEFAULT_SERVER_URL = "https://bitwarden.com";
 
+export const SENSITIVE_VALUE_PLACEHOLDER = "HIDDEN-VALUE";
+
 export const LOCAL_STORAGE_KEY = {
   PASSWORD_OPTIONS: "bw-generate-password-options",
   PASSWORD_ONE_TIME_WARNING: "bw-generate-password-warning-accepted",
@@ -12,6 +14,7 @@ export const LOCAL_STORAGE_KEY = {
   VAULT_LOCK_REASON: "vaultLockReason",
 };
 
-export const ERRORS = {
-  CLINotFound: "CLINotFound",
-} as const;
+export const VAULT_LOCK_MESSAGES = {
+  TIMEOUT: "Vault timed out due to inactivity",
+  MANUAL: "Manually locked by the user",
+};