+++ extensions/bitwarden/src/utils/preferences.ts
@@ -1,4 +1,5 @@
 import { environment, getPreferenceValues } from "@raycast/api";
+import { VAULT_TIMEOUT_MS_TO_LABEL } from "~/constants/preferences";
 import { CommandName } from "~/types/general";
 import { Preferences, TransientCopyPreferences } from "~/types/preferences";
 
@@ -23,4 +24,8 @@ export function getTransientCopyPreference(type: "password" | "other"): boolean 
   if (transientPreference === "always") return true;
   if (transientPreference === "passwords") return type === "password";
   return true;
+}
+
+export function getLabelForTimeoutPreference(timeout: string | number): string {
+  return VAULT_TIMEOUT_MS_TO_LABEL[timeout as keyof typeof VAULT_TIMEOUT_MS_TO_LABEL];
 }