+++ extensions/google-workspace/src/create-google-document.tsx
@@ -1,5 +1,5 @@
 import { createDocFromUrl } from "./helpers/docs";
 
-export default async function Command(props: { arguments: { title?: string } }) {
+export default async function Command(props: { arguments: Arguments.CreateGoogleDocument }) {
   await createDocFromUrl("document", props.arguments.title);
 }