+++ extensions/youtube/src/components/channel.tsx
@@ -15,18 +15,16 @@ import React from "react";
 import { compactNumberFormat, formatDate, getErrorMessage } from "../lib/utils";
 import { Channel, getChannel, useRefresher } from "../lib/youtubeapi";
 import { OpenChannelInBrowser, SearchChannelVideosAction, ShowRecentPlaylistVideosAction } from "./actions";
-import { addRecentChannel } from "./recent_channels";
+import { addRecentChannel, PinChannel, PinnedChannelActions, RecentChannelActions } from "./recent_channels";
 import he from "he";
 import { ViewLayout, PrimaryAction, Preferences } from "../lib/types";
 
 export function ChannelItemDetail(props: { channel: Channel; isLoading?: boolean | undefined }): JSX.Element {
   const channel = props.channel;
   let statistics;
-  let channelId: string | undefined;
   let mdParts = [];
   if (channel) {
     statistics = channel.statistics;
-    channelId = channel.id;
     const desc = channel.description || "No description";
     const title = channel.title;
     const thumbnailUrl = channel.thumbnails?.default?.url || undefined;
@@ -75,26 +73,28 @@ export function ChannelItemDetail(props: { channel: Channel; isLoading?: boolean
       actions={
         <ActionPanel>
           <ShowRecentPlaylistVideosAction
-            channel={channel}
+            channelId={channel.id}
             title="Show Recent Channel Videos"
             playlistId={channel?.relatedPlaylists?.uploads}
           />
-          <SearchChannelVideosAction channel={channel} />
-          <OpenChannelInBrowser channelId={channel.id} channel={channel} />
+          <SearchChannelVideosAction channelId={channel.id} />
+          <OpenChannelInBrowser channelId={channel.id} />
         </ActionPanel>
       }
     />
   );
 }
 
-export function ChannelItem({
-  channel,
-  actions,
-}: {
+interface ChannelItemProps {
   channel: Channel;
-  actions?: JSX.Element | undefined;
-}): JSX.Element {
+  refresh?: () => void;
+  pinned?: boolean;
+  recent?: boolean;
+}
+
+export function ChannelItem(props: ChannelItemProps): JSX.Element {
   const { view, primaryaction } = getPreferenceValues<Preferences>();
+  const { channel, refresh } = props;
   const channelId = channel.id;
   const title = he.decode(channel.title);
   let parts: string[] = [];
@@ -111,12 +111,15 @@ export function ChannelItem({
     const showDetail = (
       <Action.Push
         title="Show Details"
-        target={<ChannelItemDetail channel={channel} />}
+        target={<ChannelItemDetail {...props} />}
         icon={{ source: Icon.List, tintColor: Color.PrimaryText }}
-        onPush={() => addRecentChannel(channel)}
+        onPush={() => {
+          addRecentChannel(channel.id);
+          if (refresh) refresh();
+        }}
       />
     );
-    const openBrowser = <OpenChannelInBrowser channelId={channel.id} channel={channel} />;
+    const openBrowser = <OpenChannelInBrowser channelId={channel.id} refresh={refresh} />;
     return (
       <ActionPanel>
         <ActionPanel.Section>
@@ -133,14 +136,20 @@ export function ChannelItem({
           )}
         </ActionPanel.Section>
         <ActionPanel.Section>
-          <SearchChannelVideosAction channel={channel} />
+          <SearchChannelVideosAction channelId={channel.id} refresh={refresh} />
           <ShowRecentPlaylistVideosAction
-            channel={channel}
+            channelId={channel.id}
+            refresh={refresh}
             title="Show Recent Channel Videos"
             playlistId={channel.relatedPlaylists?.uploads}
           />
         </ActionPanel.Section>
-        {actions}
+        {props.recent && <RecentChannelActions channelId={channel.id} refresh={refresh} />}
+        {props.pinned ? (
+          <PinnedChannelActions channelId={channel.id} refresh={refresh} />
+        ) : (
+          <PinChannel channelId={channel.id} refresh={refresh} />
+        )}
       </ActionPanel>
     );
   };