+++ extensions/bitwarden/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Bitwarden Changelog
 
+## [New Command] - 2023-05-18
+
+- Add Lock Vault command
+
+## [Fix] - 2023-05-16
+
+- Fixed Generate Password (Quick) command not copying password before pasting
+
+## [Improvement] - 2023-05-12
+
+- Add different icons for each command
+
+## [Fix] - 2023-05-08
+
+- Hopefully fixed a bug that caused the extension to keep logging out right after logging in.
+
+## [Improvement] - 2023-05-04
+
+- Use session token on every bitwarden cli command
+
+## [Performance Improvement] - 2023-04-03
+
+- Implemented caching of vault items (only safe values) to speed up the Search Vault command loading time.
+
 ## [Improvement] - 2023-04-03
 
 - Display better login/unlock error messages