+++ extensions/file-manager/CHANGELOG.md
@@ -1,6 +1,15 @@
 # File Manager Changelog
 
+## [Features] - 2023-05-18
+
+- Added support for using standard shortcuts.
+
+## [Features] - 2023-05-14
+
+- Added support for renaming files and folders.
+
 ## [Fixes] - 2023-04-11
+
 - Modified behavior to prevent errors caused by broken symlinks.
 - Added an icon to indicate broken symlinks.