+++ extensions/unsplash/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unsplash Changelog
 
+## [Fix] - 2023-04-28
+
+- Fixed background updates not working.
+
 ## [Breaking Changes] - 2023-03-03
 
 - Implement OAuth to access more data.